HH-69833 move DB-related tests to example-app [an.ivanov]
