+++ example-app/src/test/java/ru/hh/nab/example/HibernateTest.java
@@ -1,4 +1,4 @@
-package ru.hh.nab.hibernate;
+package ru.hh.nab.example;
 
 import com.mchange.v2.c3p0.ComboPooledDataSource;
 import java.beans.PropertyVetoException;
