+++ example-app/src/test/resources/logback-test.xml
@@ -8,4 +8,12 @@
     <root level="info">
         <appender-ref ref="unsorted"/>
     </root>
+
+    <logger name="org.hibernate" level="debug" additivity="false">
+        <appender-ref ref="unsorted"/>
+    </logger>
+
+    <logger name="org.hsqldb" level="debug" additivity="false">
+        <appender-ref ref="unsorted"/>
+    </logger>
 </configuration>
