+++ nuts-and-bolts/pom.xml
@@ -21,12 +21,6 @@
 
     <dependencies>
         <dependency>
-            <groupId>org.hsqldb</groupId>
-            <artifactId>hsqldb</artifactId>
-            <version>2.4.0</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
             <groupId>ru.hh.metrics</groupId>
             <artifactId>hh-metrics</artifactId>
         </dependency>
@@ -275,13 +269,11 @@
         <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
-            <version>4.12</version>
             <scope>test</scope>
         </dependency>
         <dependency>
             <groupId>org.mockito</groupId>
             <artifactId>mockito-core</artifactId>
-            <version>2.8.47</version>
             <scope>test</scope>
         </dependency>
     </dependencies>
