+++ pom.xml
@@ -67,6 +67,16 @@
                 <artifactId>log4j</artifactId>
                 <version>99.0-does-not-exist</version>
             </dependency>
+            <dependency>
+                <groupId>junit</groupId>
+                <artifactId>junit</artifactId>
+                <version>4.12</version>
+            </dependency>
+            <dependency>
+                <groupId>org.mockito</groupId>
+                <artifactId>mockito-core</artifactId>
+                <version>2.8.47</version>
+            </dependency>
         </dependencies>
     </dependencyManagement>
